requirements.txt
Django==3.2.25
psycopg2-binary==2.9.9
djangorestframework==3.15.1
PyJWT==2.8.0
Pillow==9.5.0
#djangorestframework-simplejwt==5.3.1
drf-api-logger==1.1.15